util/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile util/requirements.in
#
pyyaml==6.0.2
    # via -r util/requirements.in
